notif [nfc]: Split out module notifOpen, for opening a notification
Part of the reorganization described in a recent commit.
